rpc/repe folder structure
